backend/requirements.txt
aiofiles==23.2.1
amqp==5.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
bcrypt==4.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2026.1.4
cffi==1.15.1
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.7
distro==1.9.0
dnspython==2.3.0
ecdsa==0.19.1
email-validator==2.0.0.post2
exceptiongroup==1.3.1
fastapi==0.103.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib-metadata==4.13.0
iniconfig==2.0.0
kombu==5.2.4
mock==5.2.0
openai==1.39.0
packaging==24.0
passlib==1.7.4
Pillow==9.5.0
pluggy==1.2.0
prompt_toolkit==3.0.48
pyasn1==0.4.8
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.0.3
pydantic_core==2.14.6
PyJWT==2.8.0
pymongo==4.7.3
pytest==7.4.4
pytest-asyncio==0.21.2
python-dotenv==0.21.1
python-jose==3.4.0
python-multipart==0.0.8
pytz==2025.2
redis==5.0.8
requests==2.29.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tomli==2.0.1
tqdm==4.67.1
typing_extensions==4.7.1
urllib3==1.26.20
uvicorn==0.22.0
vine==5.1.0
wcwidth==0.2.14
zipp==3.15.0
